finish refactor, add error for shunting to same thread
