+++ src/actions/feedListener/handleEpisodeNumberCommand.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import { FeedListener } from "../../../feeds/feedListener";
+import { FeedListener } from "../../listeners/feedListener/feedListener";
 
 export const handleEpisodeNumberCommand = (
   message: Message,
